[srinu]StockMangement,DeckofCardsDone

+++ JsonInventry/src/com/bridgelabz/JsonInventry/Model.java
@@ -3,21 +3,21 @@ package com.bridgelabz.JsonInventry;
 public class Model 
 {
 	private String name;
-	private float weight;
-	private float prize;
+	private double weight;
+	private double prize;
 	public String getName() {
 		return name;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
-	public float getWeight() {
+	public double getWeight() {
 		return weight;
 	}
-	public void setWeight(float weight) {
+	public void setWeight(double weight) {
 		this.weight = weight;
 	}
-	public float getPrize() {
+	public double getPrize() {
 		return prize;
 	}
 	public void setPrize(float prize) {
